화면 유지 릴리즈 빌드 산소농도 float 시리얼 서비스 변경

+++ app/src/main/java/com/coai/samin_total/Oxygen/OxygenSettingFragment.kt
@@ -53,7 +53,7 @@ class OxygenSettingFragment : Fragment() {
 
         override fun onTextChanged(s: CharSequence?, start: Int, before: Int, count: Int) {
             if (s != null && !s.toString().equals("")) {
-                selectedSensor.setMinValue = s.toString().toInt()
+                selectedSensor.setMinValue = s.toString().toFloat()
             }
         }
         override fun afterTextChanged(s: Editable?) {
